refactor: migrate legacy settings
