+++ packages/mask/shared/legacy-settings/listener.ts
