+++ packages/mask/shared/legacy-settings/settings.ts
@@ -41,8 +41,7 @@ export const dismissVerifyNextID = createComplexNetworkSettings(
 )
 export const bioDescription = createNetworkSettings('bioDescription', '')
 export const personalHomepage = createNetworkSettings('personalHomepage', '')
-// This is a misuse of concept "NetworkSettings" as "namespaced settings"
-// The refactor is tracked in https://github.com/DimensionDev/Maskbook/issues/1884
+
 /**
  * ! DO NOT use this directly to query the plugin status !
  *
@@ -51,6 +50,7 @@ export const personalHomepage = createNetworkSettings('personalHomepage', '')
  */
 // This was "currentPluginEnabled" before, but we used it to represent minimal mode now to make the settings be able to migrate.
 export const currentPluginMinimalModeNOTEnabled = createNetworkSettings('pluginsEnabled', true)
+/** @deprecated No use site. */
 export const launchPageSettings = createGlobalSettings<LaunchPage>('launchPage', LaunchPage.dashboard)
 export const currentPersonaIdentifier = createGlobalSettings('currentPersonaIdentifier', '')
