+++ packages/mask/shared/legacy-settings/types.ts
@@ -1,7 +1,6 @@
 // WARNING: Types with @public mark is exposed as public API to the native side,
 //  if you make an incompatible change in this file, it will break the API.
 import type { ChainId } from '@masknet/web3-shared-evm'
-import type { SetupGuideStep } from '../components/InjectedComponents/SetupGuide/types'
 export interface ChainBlockNumber {
     chainId: ChainId
     blockNumber: number
@@ -16,9 +15,18 @@ export type SetupGuideCrossContextStatus = {
     status?: SetupGuideStep
 }
 
-/** @public */
+/**
+ * @public
+ * @deprecated
+ */
 export enum LaunchPage {
     facebook = 'facebook',
     twitter = 'twitter',
     dashboard = 'dashboard',
+}
+export enum SetupGuideStep {
+    FindUsername = 'find-username',
+    VerifyOnNextID = 'next-id-verify',
+    PinExtension = 'pin-extension',
+    Close = 'close',
 }
