+++ packages/mask/src/UIRoot.tsx
@@ -7,7 +7,7 @@ import { getSiteType, i18NextInstance } from '@masknet/shared-base'
 import { buildInfoMarkdown } from './utils/BuildInfoMarkdown'
 import { activatedSocialNetworkUI } from './social-network'
 import { isFacebook } from './social-network-adaptor/facebook.com/base'
-import { pluginIDSettings } from './settings/settings'
+import { pluginIDSettings } from './../shared/legacy-settings/settings'
 import { getBackgroundColor } from './utils'
 import { isTwitter } from './social-network-adaptor/twitter.com/base'
 import { MaskThemeProvider } from '@masknet/theme'
