+++ packages/mask/src/components/DataSource/usePersonaConnectStatus.ts
@@ -2,9 +2,9 @@ import { DashboardRoutes, PersonaInformation } from '@masknet/shared-base'
 import stringify from 'json-stable-stringify'
 import { useMemo } from 'react'
 import Services from '../../extension/service'
-import { currentSetupGuideStatus } from '../../settings/settings'
+import { currentSetupGuideStatus } from '../../../shared/legacy-settings/settings'
 import { activatedSocialNetworkUI } from '../../social-network'
-import { SetupGuideStep } from '../InjectedComponents/SetupGuide/types'
+import { SetupGuideStep } from '../../../shared/legacy-settings/types'
 import { useLastRecognizedIdentity } from './useActivatedUI'
 import { useMyPersonas } from './useMyPersonas'
