+++ packages/mask/src/components/InjectedComponents/SetupGuide/VerifyNextID.tsx
@@ -6,8 +6,8 @@ import { VerifiedIcon } from '@masknet/icons'
 import { ActionButtonPromise } from '../../../extension/options-page/DashboardComponents/ActionButton'
 import { useWizardDialogStyles, WizardDialogProps, WizardDialog } from './WizardDialog'
 import { useFindUsernameStyles } from './FindUsername'
-import { SetupGuideStep } from './types'
-import { dismissVerifyNextID } from '../../../settings/settings'
+import { SetupGuideStep } from '../../../../shared/legacy-settings/types'
+import { dismissVerifyNextID } from '../../../../shared/legacy-settings/settings'
 import { useState } from 'react'
 import type { PersonaIdentifier } from '@masknet/shared-base'
