--- packages/mask/src/components/InjectedComponents/SetupGuide/types.ts
@@ -1,6 +1,0 @@
-export enum SetupGuideStep {
-    FindUsername = 'find-username',
-    VerifyOnNextID = 'next-id-verify',
-    PinExtension = 'pin-extension',
-    Close = 'close',
-}
