+++ packages/mask/src/extension/background-script/SocialNetworkService.ts
@@ -1,6 +1,6 @@
-import { currentSetupGuideStatus, userGuideStatus } from '../../settings/settings'
+import { currentSetupGuideStatus, userGuideStatus } from '../../../shared/legacy-settings/settings'
 import stringify from 'json-stable-stringify'
-import { SetupGuideStep } from '../../components/InjectedComponents/SetupGuide/types'
+import { SetupGuideStep } from '../../../shared/legacy-settings/types'
 import type { PersonaIdentifier, ProfileIdentifier } from '@masknet/shared-base'
 import {
     migration_in_progress_connectSite,
