+++ packages/mask/src/plugins/Pets/settings.ts
@@ -1,4 +1,4 @@
-import { createGlobalSettings } from '../../settings/createSettings'
+import { createGlobalSettings } from '../../../shared/legacy-settings/createSettings'
 import { PetsPluginID } from './constants'
 
 export const petShowSettings = createGlobalSettings(`${PetsPluginID}+selectedClosePet`, true)
