+++ packages/mask/src/social-network-adaptor/twitter.com/injection/PostDialogHint.tsx
@@ -8,7 +8,7 @@ import { startWatch } from '../../../utils/watcher'
 import { makeStyles, MaskColorVar } from '@masknet/theme'
 import { alpha } from '@mui/material'
 import { twitterBase } from '../base'
-import { sayHelloShowed } from '../../../settings/settings'
+import { sayHelloShowed } from '../../../../shared/legacy-settings/settings'
 import { makeTypedMessageText } from '@masknet/typed-message'
 import { useI18N } from '../../../utils'
 import { hasEditor, isCompose } from '../utils/postBox'
