+++ packages/mask/src/utils/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from './useControlledDialog'
 export * from './useMatchXS'
 export * from './useQueryNavigatorPermission'
-export * from './useSettingSwitcher'
+export * from './useSwitcher'
 export * from './useSuspense'
 export * from './useLocationChange'
 export * from './useNextIDWallets'
